switch fvalue to float, fix compile issues and warnings steming from change in sizeof(fvalue)

+++ eisgenerator/eistype.h
@@ -5,7 +5,7 @@
 #include <cmath>
 #include <filesystem>
 
-typedef double fvalue;
+typedef float fvalue;
 
 namespace eis
 {
